requirements.txt
plotly
kaleido
requests
pandas
python-dotenv